+++ html/desafioIntegrador.html
@@ -41,8 +41,9 @@
         aqui vai o resultado...
       </div>
     </div>
-  </main>
-
-  <script src="../desafioIntegrador.js"></script> 
+  </main>  
+  
+  <script src="../js/dados.js"></script>
+  <script type="module"  src="../js/desafioIntegrador.js"></script>
 </body>
 </html>
